requirements.txt
asgiref==3.4.1
attrs==21.2.0
black==21.8b0
click==8.0.1
colorama==0.4.4
Django==3.2.7
django-admin-honeypot==1.1.0
django-cleanup==5.2.0
django-cors-headers==3.8.0
django-environ==0.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
inflection==0.5.1
jsonschema==3.2.0
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.3.2
platformdirs==2.3.0
psycopg2==2.9.1
PyJWT==2.1.0
pyrsistent==0.18.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.28
six==1.16.0
sqlparse==0.4.2
tomli==1.2.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
